+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## Unreleased
 
+### Added
+- Added `benf\neo\elements\db\BlockQuery::findStructureId()`
+- Added `benf\neo\elements\db\BlockQuery::$findStructureId`
+
 ### Fixed
 - Fixed an error that occurred when calling `benf\neo\Field::getBlockTypeFields()`
+- Fixed a bug where nested Matrix entries could not be created in new Neo blocks until after saving the owner element
 
 ## 5.0.4 - 2024-06-26